fix: mac releated

+++ deploy/packer.py
@@ -29,7 +29,6 @@ else:
     tmp_p = prog_path.parent.joinpath("temp")
 path = prog_path.parent
 sys.path.append(str(prog_path))
-from utils.docs import MarkdownConverter, MdHtml
 
 api_github = "https://api.github.com"
 github_token = "**create token by your github account**"
@@ -145,12 +144,15 @@ class Packer(Proj):
             specified = ('runtime', '_pystand_static.int',
                          f'{self._proj}.exe')
             mode = "w"
-        else:
+        elif self.preset_zip_file.exists():
             shutil.copy(self.preset_zip_file, self.zip_file)
+        # 仅 CI 环境，处理 【过滤.git后的script、version.json、CGS.bat】 进经 workflow 安装了 runtime 全部依赖的 perset 包
         with py7zr.SevenZipFile(zip_file, mode, filters=[{"id": py7zr.FILTER_LZMA2}]) as zip_f:
             for file in tqdm(tuple(specified)):
                 if file == f'{Proj.proj}.bat':
                     zip_f.write(path.joinpath(f"scripts/deploy/launcher/{file}"), arcname=file)
+                elif file == 'version.json':
+                    zip_f.write(path.joinpath(f"scripts/{file}"), arcname=f"scripts/{file}")
                 elif path.joinpath(file).exists():
                     zip_f.writeall(file)
         if not self.zip_file.exists():
@@ -162,22 +164,14 @@ class PackerMacOS(Packer):
     zip_file = path.joinpath(f'{_proj}.7z')
     preset_zip_file = tmp_p.joinpath(f'{_proj}_preset.7z')
     scripts_path = "CGS.app/Contents/Resources/scripts"
-    tmp_guide_html = tmp_p.joinpath('CGS_macOS_first_guide.html')
 
     def __init__(self, ver):
         super(PackerMacOS, self).__init__(tuple(), ver)
 
-    def set_guide(self):
-        _html = """<!doctype html><html lang="en"><body><h1><p>查看 <a href="https://jasoneri.github.io/ComicGUISpider/deploy/mac-required-reading">mac部署</a></p></h1></body></html>"""
-        with open(self.tmp_guide_html, 'w', encoding='utf-8') as f:
-            f.write(_html)
-
     def pre_packup(self):
-        """ 1. 生成guide.html
-            2. 预处理CGS.app的结构
-            3. dos2unix处理项目文本文件"""
+        """ 1. 预处理CGS.app的结构
+            2. dos2unix处理项目文本文件"""
         super(PackerMacOS, self).pre_packup()
-        self.set_guide()
         mac_7z_p = path.joinpath(self.scripts_path.rsplit("/", maxsplit=1)[0])
         mac_7z_p.mkdir(parents=True, exist_ok=True)
         mac_scripts_path = mac_7z_p.joinpath("scripts")
@@ -206,7 +200,6 @@ class PackerMacOS(Packer):
             for file in tqdm(tuple(specified)):
                 if path.joinpath(file).exists():
                     zip_f.writeall(path.joinpath(file), arcname=file)
-            zip_f.write(self.tmp_guide_html, arcname='CGS_macOS_first_guide.html')
         if not self.zip_file.exists():
             self.packup()
 
@@ -222,7 +215,7 @@ def packup_windows(ver):
     # # clean()
     Clean.end_work(path.joinpath("scripts").rglob("__pycache__"), path.joinpath("runtime").rglob("__pycache__"),
         (path.joinpath("scripts/log"), path.joinpath("scripts/deploy/gitee_t.json")))  # step 0 必清runtime cache，太大了
-    packer = Packer(('scripts', f'{Proj.proj}.bat'), ver=ver)
+    packer = Packer(('scripts', 'version.json', f'{Proj.proj}.bat'), ver=ver)
     packer.packup()  # step 2
     # # Clean.end_work(('CGS.7z',))  # step 4
     # # If error occur, exegesis previous step and run again
